# docs/swagger.yaml
basePath: /
definitions:
  domain.AssignUserToGroupRequest:
    properties:
      group_id:
        type: string
    required:
    - group_id
    type: object
...
  domain.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        type: string
    type: object
  domain.Group:
    properties:
      created_at:
        type: string
      id:
        type: string
      manager_id:
        type: string
      name:
        type: string
      org_id:
        type: string
      shift_id:
        type: string
    required:
    - name
    type: object
...
  domain.InviteEmployeeRequest:
    properties:
      email:
        type: string
      group_id:
        type: string
      role:
        enum:
        - MANAGER
        - EMPLOYEE
        type: string
    required:
    - email
    - role
    type: object
...
  domain.Organization:
    properties:
      created_at:
        type: string
      default_location_lat:
        type: number
      default_location_long:
        type: number
      email:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - default_location_lat
    - default_location_long
    - email
    - name
    type: object
  domain.OrganizationMemberDetail:
    properties:
      created_at:
        type: string
      group_id:
        type: string
      id:
        type: string
      org_id:
        type: string
      role:
        description: OWNER, MANAGER, EMPLOYEE
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: string
    type: object
...
  domain.Shift:
    properties:
      allowed_late_minutes:
        type: integer
      created_at:
        type: string
      end_time:
        description: HH:MM
        type: string
      id:
        type: string
      name:
        type: string
      org_id:
        type: string
      start_time:
        description: HH:MM
        type: string
      timezone:
        type: string
      working_days:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - end_time
    - name
    - start_time
    - timezone
    - working_days
    type: object
...
  domain.UpdateEmployeeRequest:
    properties:
      group_id:
        type: string
      role:
        enum:
        - MANAGER
        - EMPLOYEE
        type: string
    required:
    - role
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone_number:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for Check-In/Out Service
  title: Check-In Service API
  version: "1.0"
paths:
...
  /organizations:
    get:
      consumes:
      - application/json
      description: List organizations where the user is a member
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Organization'
            type: array
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organizations
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: Create a new organization and assign the creator as OWNER
      parameters:
      - description: Organization Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Organization'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Organization'
        "400":
          description: invalid request body or validation errors
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new organization
      tags:
      - Organization
  /organizations/{org_id}:
    delete:
      consumes:
      - application/json
      description: Delete an organization (Owner only)
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - Organization
    get:
      consumes:
      - application/json
      description: Get details of a specific organization
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Organization'
        "404":
          description: organization not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get organization details
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: Update details of an organization (Owner only)
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Organization Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Organization'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: invalid request body or validation errors
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update organization details
      tags:
      - Organization
  /organizations/{org_id}/employees:
    get:
      consumes:
      - application/json
      description: Get list of employees in the organization (Owner/Manager only)
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.OrganizationMemberDetail'
            type: array
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employees
      tags:
      - Organization
  /organizations/{org_id}/employees/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove an employee from the organization (Owner/Manager only)
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove employee
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: Update employee role or group (Owner/Manager only)
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: invalid request body or validation errors
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update employee
      tags:
      - Organization
  /organizations/{org_id}/groups:
    post:
      consumes:
      - application/json
      description: Create a new group/team for the organization
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Group Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Group'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Group'
        "400":
          description: invalid request body or validation errors
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a group
      tags:
      - Organization
  /organizations/{org_id}/invitations:
    post:
      consumes:
      - application/json
      description: Invite an employee to the organization
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Invite Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.InviteEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: invalid request body or validation errors
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite an employee
      tags:
      - Organization
  /organizations/{org_id}/members/{user_id}:
    put:
      consumes:
      - application/json
      description: Assign a user to a specific group within the organization
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Assign Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.AssignUserToGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: invalid request body or validation errors
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign user to group
      tags:
      - Organization
...
  /organizations/{org_id}/shifts:
    post:
      consumes:
      - application/json
      description: Create a new shift for the organization
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Shift Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Shift'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Shift'
        "400":
          description: invalid request body or validation errors
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a shift
      tags:
      - Organization
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"